requirements.txt
anthropic>=0.20.0
pydantic>=2.0.0
rich>=13.0.0
gitpython>=3.1.0
pillow>=10.0.0
mcp>=1.0.0
tiktoken
openai>=1.0.0
transformers>=4.0.0
